+++ DoAnPlatformer/Assets/Scripts/Mob/enemyHealth.cs
@@ -11,22 +11,20 @@ public class enemyHealth : MonoBehaviour,IDamageable
     private Animator anim;
     private bool dead;
 
-    [Header("iFrames")]
-    [SerializeField] private float iFramesDuration;
-    [SerializeField] private int numberOfFlashes;
-    private SpriteRenderer spriteRend;
     
     void Start()
     {
         currentHealth = startingHealth;
         anim = GetComponent<Animator>();
-        spriteRend = GetComponent<SpriteRenderer>();  
     }
 
     public void TakeDamage(float amount)
     {
         currentHealth = Mathf.Max(currentHealth - amount, 0.0f);
-       
+       if(currentHealth > 0)
+        {
+            
+        }
         // if health reach to zero we call the die function
         if (currentHealth == 0)
         {
@@ -38,17 +36,5 @@ public class enemyHealth : MonoBehaviour,IDamageable
         Debug.Log("Enemy is Dead");
         anim.SetTrigger("die");
         Destroy(gameObject);
-    }
-    private IEnumerator Invunerability()
-    {
-        Physics2D.IgnoreLayerCollision(3, 8, true);
-        for (int i = 0; i < numberOfFlashes; i++)
-        {
-            spriteRend.color = new Color(1, 0, 0, 0.5f);
-            yield return new WaitForSeconds(iFramesDuration / (numberOfFlashes * 2));
-            spriteRend.color = Color.white;
-            yield return new WaitForSeconds(iFramesDuration / (numberOfFlashes * 2));
-        }
-        Physics2D.IgnoreLayerCollision(3, 8, false);
     }
 }
